i tried to get around this debugging stop, but i think it's inevitable, will  need these to figure out WTF is going on internally.
git-svn-id: svn://okws.org/okws2/devel/3.0@4897 7d287422-6eea-da11-b677-00123f2a783a

+++ libpub/okdbg-int.h
@@ -65,7 +65,7 @@ public:
   {
     vec<str> v;
     okdbg_dump_vec (&v);
-    for (u_int i = 0; i < v.size () ; i++) {
+    for (size_t i = 0; i < v.size () ; i++) {
       okdbg_warn (l, v[i]);
     }
   }
